SPACE between QUOTE
