+++ src/builtins/export.c
@@ -6,7 +6,7 @@
 /*   By: hho-troc <hho-troc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/09 14:44:38 by hho-troc          #+#    #+#             */
-/*   Updated: 2025/05/30 12:38:59 by hho-troc         ###   ########.fr       */
+/*   Updated: 2025/06/01 10:31:45 by hho-troc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
