+++ src/parsing/join_tokens_arg.c
@@ -6,25 +6,15 @@
 /*   By: hho-troc <hho-troc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/29 19:08:42 by hho-troc          #+#    #+#             */
-/*   Updated: 2025/05/31 12:41:43 by hho-troc         ###   ########.fr       */
+/*   Updated: 2025/06/01 11:36:37 by hho-troc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-// static bool	should_break(char *arg, t_token *tok, t_quote prev_quote)
-// {
-// 	if (arg[0] != '\0' && !tok->glued && prev_quote != Q_NONE)
-// 		return (true);
-// 	if (arg[0] != '\0' && !tok->glued)
-// 		return (true);
-// 	return (false);
-// }
-
 static bool	should_break(char *arg, t_token *tok, t_quote prev_quote)
 {
-	(void)prev_quote; // prev_quote is not used in this function
-	
+	(void)prev_quote;
 	if (arg[0] == '\0')
 		return (false);
 	if (tok->glued)
@@ -32,8 +22,6 @@ static bool	should_break(char *arg, t_token *tok, t_quote prev_quote)
 	if (ft_strcmp(arg, "$") == 0 && tok->quote_type == Q_NONE && \
 												ft_isalpha(tok->str[0]))
 		return (true);
-	// if (prev_quote != Q_NONE || tok->quote_type != Q_NONE)
-	// 	return (false);
 	return (true);
 }
