+++ src/parsing/parse_pipeline.c
@@ -6,7 +6,7 @@
 /*   By: hho-troc <hho-troc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/23 14:14:49 by hho-troc          #+#    #+#             */
-/*   Updated: 2025/05/23 12:47:09 by hho-troc         ###   ########.fr       */
+/*   Updated: 2025/06/01 11:37:23 by hho-troc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
