+++ src/tokenizing/token_utils.c
@@ -6,26 +6,11 @@
 /*   By: hho-troc <hho-troc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/25 12:37:11 by hho-troc          #+#    #+#             */
-/*   Updated: 2025/05/31 12:02:13 by hho-troc         ###   ########.fr       */
+/*   Updated: 2025/06/01 11:11:22 by hho-troc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-// t_node	get_token_type(const char *str)
-// {
-// 	if (!ft_strcmp(str, "|"))
-// 		return (PIPE);
-// 	if (!ft_strcmp(str, ">"))
-// 		return (R_OUT);
-// 	if (!ft_strcmp(str, "<"))
-// 		return (R_IN);
-// 	if (!ft_strcmp(str, ">>"))
-// 		return (R_A);
-// 	if (!ft_strcmp(str, "<<"))
-// 		return (HD);
-// 	return (CMD);
-// }
 
 t_node	get_token_type(const char *str, t_quote qt)
 {
@@ -45,22 +30,6 @@ t_node	get_token_type(const char *str, t_quote qt)
 	return (CMD);
 }
 
-
-// t_token	*create_t(char *str, t_quote quote_type)
-// {
-// 	t_token	*new;
-
-// 	new = malloc(sizeof(t_token));
-// 	if (!new)
-// 		return (NULL);
-// 	new->str = str;
-// 	new->type = get_token_type(str);
-// 	new->quote_type = quote_type;
-// 	new->is_dollar_quote = false;
-// 	new->next = NULL;
-// 	return (new);
-// }
-
 t_token	*create_t(char *str, t_quote quote_type)
 {
 	t_token	*new;
@@ -69,7 +38,7 @@ t_token	*create_t(char *str, t_quote quote_type)
 	if (!new)
 		return (NULL);
 	new->str = str;
-	new->type = get_token_type(str, quote_type); // 這裡有差
+	new->type = get_token_type(str, quote_type);
 	new->quote_type = quote_type;
 	new->is_dollar_quote = false;
 	new->next = NULL;
